pyproject.toml
[project]
name = "redis-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "redis[hiredis]>=7.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
